% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{alignment_to_dnabin}
\alias{alignment_to_dnabin}
\title{Convert alignment to DNA bin}
\usage{
alignment_to_dnabin(path)
}
\arguments{
\item{path}{The alignment object.}
}
\value{
DNA bin
}
\description{
Convert alignment to DNA bin
}
\examples{
alignment_to_dnabin("./alignment.fas")
}
