% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{build_max_parsimony}
\alias{build_max_parsimony}
\title{Build phylogenetic tree using Maximum Parismony method}
\usage{
build_max_parsimony(alignment, outgroup = NULL, seed = 123, directory = ".")
}
\arguments{
\item{alignment}{The alignment object.}

\item{outgroup}{The sample name to serve as the outgroup.}

\item{seed}{The value for bootstrapping.}

\item{directory}{The directory to save the png plot.}
}
\value{
A phylogenetic tree plot.
}
\description{
Build phylogenetic tree using Maximum Parismony method
}
\examples{
build_max_parsimony(my_alignment, seed = 1000)
}
