% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{build_ml_tree}
\alias{build_ml_tree}
\title{Build phylogenetic tree using Maximum Likelihood method}
\usage{
build_ml_tree(
  alignment,
  outgroup = NULL,
  seed = 123,
  bs_reps = 100,
  directory = "."
)
}
\arguments{
\item{alignment}{The alignment object.}

\item{outgroup}{The sample name to serve as the outgroup.}

\item{seed}{The value for bootstrapping.}

\item{bs_reps}{The number of boostrap replicates to run.}

\item{directory}{The directory to save the png plot.}
}
\value{
A list specifying the (1) model and (2) phylogenetic tree plot.
}
\description{
Build phylogenetic tree using Maximum Likelihood method
}
\examples{
build_ml_tree(my_alignment, seed = 1000, bs_reps = 1000)
}
