% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{build_nj_tree}
\alias{build_nj_tree}
\title{Build phylogenetic tree using NJ method}
\usage{
build_nj_tree(alignment, outgroup = NULL, seed = 123, model = model)
}
\arguments{
\item{alignment}{The alignment object.}

\item{outgroup}{The sample name to serve as the outgroup.}

\item{seed}{The value for bootstrapping.}

\item{model}{The substitution model.}
}
\value{
A phylogenetic tree plot.
}
\description{
Build phylogenetic tree using NJ method
}
