% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_extraction.R
\name{calc_concordance}
\alias{calc_concordance}
\title{Match samples for concordance analysis}
\usage{
calc_concordance(file1, file2, phased = FALSE)
}
\arguments{
\item{file1}{The file path of sequencing results using technique 1. Expected sample and genotype information.}

\item{file2}{The file path of sequencing results using technique 2. Expected sample and genotype information.}

\item{phased}{Indicates if genotypes are phased. Default is FALSE.}
}
\value{
The data frame containing the genotype of samples present in both file1 and file2.
}
\description{
Match samples for concordance analysis
}
\examples{
calc_concordance(file1 = "ngs.csv", file2 = 'wgs.xlsx')
}
