% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_statistics.R
\name{calc_genotype_freq}
\alias{calc_genotype_freq}
\title{Calculate the genotype frequencies per population}
\usage{
calc_genotype_freq(df, pop = NULL)
}
\arguments{
\item{df}{The dataframe containing sample, population, and genotype information.}

\item{pop}{The total number of samples in the data. Required if employing five-event minimum allele frequency.}
}
\value{
A list containing genotype frequency (overall and by population).
}
\description{
Calculate the genotype frequencies per population
}
\examples{
calc_genotype_freq(allele_freq)
}
