% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_statistics.R
\name{calc_iisnps_params}
\alias{calc_iisnps_params}
\title{Calculate forensic parameters for iisnps}
\usage{
calc_iisnps_params(geno_freqs, profile = NULL, theta = 0)
}
\arguments{
\item{geno_freqs}{The dataframe of genotype frequencies per population.}

\item{profile}{(optional) If target profile will be provided for identification.}

\item{theta}{The coefficient of ancestry or inbreeding to adjust calculation of match probabilities. Required if profile is not NULL.}
}
\value{
A list of dataframes containing results.
}
\description{
Calculate forensic parameters for iisnps
}
\examples{
calc_iisnps_params(geno_freq)
}
