% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{calc_msa}
\alias{calc_msa}
\title{Perform multiple sequence alignment}
\usage{
calc_msa(files, algorithm)
}
\arguments{
\item{files}{The DNA bin object.}

\item{algorithm}{The method for alignment (ClustalW, ClustalOmega, MUSCLE).}
}
\value{
A list of aligned sequences.
}
\description{
Perform multiple sequence alignment
}
\examples{
calc_msa(fasta_files, algorithm = "ClustalW")
}
