% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_classification.R
\name{calculate_naive_bayes}
\alias{calculate_naive_bayes}
\title{Perform Naive Bayes classification on forensic SNPs}
\usage{
calculate_naive_bayes(file)
}
\arguments{
\item{file}{The file path to file with sample, population, and genotype information.}
}
\value{
A list containing the prediction rate and other model metrices.
}
\description{
Perform Naive Bayes classification on forensic SNPs
}
\examples{
calculate_naive_bayes("./file/my_pop_file.xlsx")
}
