% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{clean_input_data}
\alias{clean_input_data}
\title{Homogenize data format}
\usage{
clean_input_data(file)
}
\arguments{
\item{input}{The input as a dataframe.}
}
\value{
A dataframe containing information from the input file.
}
\description{
Homogenize data format
}
\examples{
clean_input_data(my_dataframe)
}
