% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_statistics.R
\name{compute_fst}
\alias{compute_fst}
\title{Compute FST}
\usage{
compute_fst(fsnps_gen)
}
\arguments{
\item{fsnps_gen}{The genind data containing genotype and population information.}
}
\value{
A list of FST stats results.
}
\description{
Compute FST
}
\examples{
compute_fst(fsnps_gen = my_genind)
}
