% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_statistics.R
\name{compute_pop_stats}
\alias{compute_pop_stats}
\title{Compute heterozygosities, MAR, and inbreeding}
\usage{
compute_pop_stats(fsnps_gen)
}
\arguments{
\item{fsnps_gen}{The genind data containing genotype and population information.}
}
\value{
A list containing the metrices.
}
\description{
Compute heterozygosities, MAR, and inbreeding
}
\examples{
compute_pop_stats(fsnps_gen = my_genind)
}
