% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{convert_from_plink2}
\alias{convert_from_plink2}
\title{Convert PLINK 1.9 files to other formats}
\usage{
convert_from_plink2(prefix, output_type, output.dir = ".", ref = NULL)
}
\arguments{
\item{prefix}{The prefix of the PLINK files.}

\item{output_type}{The desired output file type.}

\item{output.dir}{The directory to save file. Default is the working directory.}

\item{ref}{The population metadata of samples. Required only if output_type is CSV.}
}
\value{
The file path of the output.
}
\description{
Convert PLINK 1.9 files to other formats
}
\examples{
convert_from_plink2(prefix = "my_plink_files", output_type = "vcf2", output.dir = "./unpacked")
}
