% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{convert_to_plink}
\alias{convert_to_plink}
\title{Convert files to PLINK1.9 files}
\usage{
convert_to_plink(input.file, name = "converted_to_plink")
}
\arguments{
\item{input.file}{The filepath of the dataset (either VCF, VCF.GZ, BCF, or PLINK1.9 files).}

\item{name}{The prefix of the output file.}

\item{output.dir}{The filename of the input file, NULL if PLINK files.}

\item{isplink}{Indicates whether input file/s are PLINK files. Default is FALSE.}

\item{output_chr}{The chromosome formatting for output files based on PLINK2.0. Default is '26' indicating numeric codes.}
}
\value{
The prefix of PLINK2.0 files.
}
\description{
Convert files to PLINK1.9 files
}
