% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{convert_to_plink2}
\alias{convert_to_plink2}
\title{Convert files to PLINK2.0 files}
\usage{
convert_to_plink2(
  input.file,
  original_name = NULL,
  isplink = FALSE,
  name = "convert_to_plink2",
  output_chr = "26"
)
}
\arguments{
\item{input.file}{The filepath of the dataset (either VCF, VCF.GZ, BCF, or PLINK1.9 files).}

\item{original_name}{The filename of the input file, NULL if PLINK files.}

\item{isplink}{Indicates whether input file/s are PLINK files. Default is FALSE.}

\item{name}{The prefix of the output file.}

\item{output_chr}{The chromosome formatting for output files based on PLINK2.0. Default is '26' indicating numeric codes.}
}
\value{
The prefix of PLINK2.0 files.
}
\description{
Convert files to PLINK2.0 files
}
\examples{
convert_to_plink2('my_vcf.vcf', original_name = "my_vcf", isplink = FALSE, plink_path = "./plink2/plink2.exe, name = "converted_file", output_chr = "26")
convert_to_plink2('my_vcf.vcf', original_name = "my_vcf", isplink = FALSE, plink_path = "./plink2/plink2.exe, name = "converted_file", output_chr = "MT")
}
