% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_extraction.R
\name{create_range_file}
\alias{create_range_file}
\title{Create POS range file for merging}
\usage{
create_range_file(pos_input, addID = FALSE, output_dir = ".")
}
\arguments{
\item{pos_input}{The path to file containing marker information. Required columns are 1. rsID, 2. chromosome, 3. position if addID is TRUE. Else, only the last two columns are expected.}

\item{addID}{Indicate whether rsIDs will be added to the extracted SNPs.}

\item{output_dir}{The directory to save the range files.}
}
\value{
A list of file paths of range files.
}
\description{
Creates a range file necessary when extracting SNPs by base pair position.
The rsIDs are assumed to be absent from input file. The function also creates a file for adding rsID if addID set to TRUE.
Default is FALSE.
}
\keyword{internal}
