% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{csv_to_gentibble}
\alias{csv_to_gentibble}
\title{Convert file to gen_tibble object}
\usage{
csv_to_gentibble(file, loci.meta = loci.meta)
}
\arguments{
\item{file}{The dataframe of SNP calls.}

\item{loci.meta}{The file path to the reference file containing metadata of markers with rsID, chr, pos, genetic distance, ref allele, and alt allele information.}
}
\value{
A gentibble object.
}
\description{
Convert file to gen_tibble object
}
\examples{
csv_to_gentibble('snp_calls.csv', loci.meta = "markers_meta.xlsx")
}
