% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_statistics.R
\name{depth_from_vcf}
\alias{depth_from_vcf}
\title{Plot Depth of Coverage}
\usage{
depth_from_vcf(
  vcf,
  output.dir = ".",
  reference,
  palette = NULL,
  width = 10,
  height = 8,
  dpi = 300
)
}
\arguments{
\item{vcf}{The file path of the VCF input file.}

\item{output.dir}{The directory to save the plots. Default is the working directory.}

\item{reference}{The file path to the population data of the samples.}

\item{palette}{The color to differentiate the samples.}

\item{width}{The width of the png output.}

\item{height}{The height of the png output.}

\item{dpi}{The quality of the png output.}
}
\value{
Depth plot.
}
\description{
Plot Depth of Coverage
}
\examples{
depth_from_vcf("my_file.vcf", reference = "pop_data.xlsx", palette = "Set1")
}
