% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_structure_analysis.R
\name{.structureParseQmat}
\alias{.structureParseQmat}
\title{Parse matrix files
Sections commented out were revised.}
\usage{
.structureParseQmat(q.mat.txt, pops)
}
\description{
Parse matrix files
Sections commented out were revised.
}
\keyword{internal}
