% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_statistics.R
\name{evaluate_file}
\alias{evaluate_file}
\title{Evaluate if input file is an allele frequency or gt frequency table}
\usage{
evaluate_file(df, sample_size = 50, genotype = "^[A-Z]/[A-Z]$")
}
\arguments{
\item{df}{The genotype input as a dataframe.}

\item{sample_size}{The total number of samples in the file.}

\item{genotype}{The format of genotype data.}
}
\value{
A string indicating the data format
}
\description{
Evaluate if input file is an allele frequency or gt frequency table
}
\keyword{internal}
