% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_statistics.R
\name{export_pop_results}
\alias{export_pop_results}
\title{Compile and export calculated population summary statistics}
\usage{
export_pop_results(
  allele_freq,
  priv_alleles,
  stats_matrix,
  hw_matrix,
  fst_matrix,
  dir = tempdir()
)
}
\arguments{
\item{allele_freq}{The dataframe of calculated allele frequencies.}

\item{priv_alleles}{A list/string/dataframe of the total number of calculated private alleles per population.}

\item{stats_matrix}{A list of the calculated heterozygosities, MAR, and inbreeding coefficient.}

\item{hw_matrix}{A list of the HWE test results.}

\item{fst_matrix}{A list of the FST metrices.}

\item{dir}{The directory to save the result.}
}
\value{
An excel (.xlsx) file containing the results in individual sheets.
}
\description{
Compile and export calculated population summary statistics
}
\examples{
export_pop_results(allele_freq, priv_alleles, stats_matrix, hw_matrix, fst_matrix, dir = ".")
}
