% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_extraction.R
\name{extract_by_ID_pgen}
\alias{extract_by_ID_pgen}
\title{Extract SNPs by rsID}
\usage{
extract_by_ID_pgen(
  pgen_prefix,
  snps_list,
  output_dir = ".",
  merged_file = "extracted_file"
)
}
\arguments{
\item{pgen_prefix}{The prefix of the PLINK2.0 files.}

\item{snps_list}{The dataframe of the list of SNPs to be extracted.}

\item{output_dir}{The directory to save output file. Default is working directory.}

\item{merged_file}{The name of the output file.}
}
\value{
The file path of result.
}
\description{
Extract SNPs by rsID
}
\examples{
extract_by_ID_pgen(pgen_prefix = 'my_file', snps_list = 'list_of_snps')
}
