% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_extraction.R
\name{extract_by_pos_pgen}
\alias{extract_by_pos_pgen}
\title{Extract SNPs by GRCh38/37 positions}
\usage{
extract_by_pos_pgen(
  pos_list,
  pgen_prefix,
  output_dir = ".",
  merged_file = "extracted_file"
)
}
\arguments{
\item{pos_list}{The file path of the list of SNPs to be extracted.}

\item{pgen_prefix}{The prefix of the PLINK2.0 files.}

\item{output_dir}{The directory to save output file. Default is working directory.}

\item{merged_file}{The name of the output file.}
}
\value{
The file path of extracted file.
}
\description{
Extract SNPs by GRCh38/37 positions
}
\examples{
extract_by_pos_pgen(pos_list = "my_snps.xlsx", pgen_prefix = 'my_file', merged_file = "extracted_file")
}
