% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_extraction.R
\name{extract_POStoID_pgen}
\alias{extract_POStoID_pgen}
\title{Extract SNPs by GRCh38/37 positions and add rsID}
\usage{
extract_POStoID_pgen(pos_list, pgen_prefix, output_dir = ".")
}
\arguments{
\item{pos_list}{The file path of the list of SNPs to be extracted.}

\item{pgen_prefix}{The prefix of the PLINK2.0 files.}

\item{output_dir}{The directory to save output file. Default is working directory.}
}
\value{
The file path of result.
}
\description{
Extract SNPs by GRCh38/37 positions and add rsID
}
\examples{
extract_POStoID_pgen(pos_list = "my_snps.xlsx", pgen_prefix = 'my_file')
}
