% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_structure_analysis.R
\name{genind2structure2}
\alias{genind2structure2}
\title{Genind to STRUCTURE}
\usage{
genind2structure2(data, file = "", pops = TRUE, markers = TRUE, unix = FALSE)
}
\arguments{
\item{input}{The input as a dataframe.}
}
\value{
A dataframe containing information from the input file.
}
\description{
Genind to STRUCTURE
}
\keyword{internal}
