% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{get_plink2_path}
\alias{get_plink2_path}
\title{Get PLINK2 executable path}
\usage{
get_plink2_path()
}
\value{
The file path for PLINK 2.0.
}
\description{
Get PLINK2 executable path
}
\keyword{internal}
