% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{get_plink_path}
\alias{get_plink_path}
\title{Get PLINK1.9 executable path}
\usage{
get_plink_path()
}
\value{
The file path for PLINK 1.9.
}
\description{
Get PLINK1.9 executable path
}
\keyword{internal}
