% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{load_csv_xlsx_files}
\alias{load_csv_xlsx_files}
\title{Read CSV/XLSX/TXT files}
\usage{
load_csv_xlsx_files(input)
}
\arguments{
\item{input}{The file path of input.}
}
\value{
A dataframe containing information from the input file.
}
\description{
Read CSV/XLSX/TXT files
}
\keyword{internal}
