% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{load_vcf_files}
\alias{load_vcf_files}
\title{Read VCF file and convert to DF object}
\usage{
load_vcf_files(vcf, output.dir = ".")
}
\arguments{
\item{vcf}{The VCF file path.}

\item{output.dir}{The directory to save the unpacked and merged files if files are zipped. Default is the working directory.}
}
\value{
The genotype dataframe with sample ID.
}
\description{
Read VCF file and convert to DF object
}
\examples{
load_vcf_files('my_file.vcf')
}
