% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_extraction.R
\name{plot_concordance}
\alias{plot_concordance}
\title{Perform concordance analysis}
\usage{
plot_concordance(dataframe)
}
\arguments{
\item{dataframe}{The dataframe containing genotype information of samples sequenced using two different techniques.}
}
\value{
A list containing the dataframe of the tally of concordant and discordant calls and the concordance plot.
}
\description{
Compares genetic data information generated by different sequencing techniques.
It compares the genetic data of the same individual from technique 1 and technique 2.
}
\examples{
plot_concordance(dataframe = "merged_samples")
}
