% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_statistics.R
\name{plot_fst}
\alias{plot_fst}
\title{Plot FST across populations}
\usage{
plot_fst(fst_df, out_dir)
}
\arguments{
\item{out_dir}{The directory to save the plot.}

\item{Het_fsnps_df}{A dataframe containing the calculated FST values.}
}
\value{
PNG file of the heatmap.
}
\description{
Plot FST across populations
}
\keyword{internal}
