% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_statistics.R
\name{plot_heterozygosity}
\alias{plot_heterozygosity}
\title{Plot heterozygosity tables}
\usage{
plot_heterozygosity(Het_fsnps_df, out_dir)
}
\arguments{
\item{Het_fsnps_df}{A dataframe containing the observed and expected heterozygosities ("Ho" and "He").}

\item{out_dir}{The directory to save the plot.}
}
\value{
Plot of observed vs expected heterozygosity per population.
}
\description{
Plot heterozygosity tables
}
\keyword{internal}
