% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_statistics.R
\name{pop_breakdown}
\alias{pop_breakdown}
\title{Calculate the population breakdown of samples}
\usage{
pop_breakdown(file, column)
}
\arguments{
\item{file}{The input file as dataframe.}

\item{column}{(string) The basis for grouping samples, it should exist as a column in the dataframe.}
}
\value{
The dataframe of tally count.
}
\description{
Calculate the population breakdown of samples
}
\examples{
pop_breakdown(info_with_meta, column = "Population")
}
