% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{prepare_input_dataset}
\alias{prepare_input_dataset}
\title{Handle file conversion to PLINK}
\usage{
prepare_input_dataset(input_file, output.dir = ".")
}
\arguments{
\item{input_file}{The filepath of the zipped file.}

\item{output.dir}{The directory to save the unpacked files. Default is the working directory.}
}
\value{
The prefix of the converted dataset.
}
\description{
Handle file conversion to PLINK
}
\examples{
prepare_input_dataset('my_zipped_file.zip', output.dir = "./unpacked")
}
