% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{prepare_input_dataset_archive}
\alias{prepare_input_dataset_archive}
\title{Convert zipped files to PLINK 2.0 files}
\usage{
prepare_input_dataset_archive(input_file, output.dir = ".")
}
\arguments{
\item{input_file}{The filepath of the zipped file.}

\item{output.dir}{The directory to save the unpacked files. Default is the working directory.}
}
\value{
The prefix of the merged dataset.
}
\description{
Convert zipped files to PLINK 2.0 files
}
\examples{
prepare_input_dataset_archive('my_zipped_file.zip', output.dir = "./unpacked")
}
