% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_statistics.R
\name{q_matrices}
\alias{q_matrices}
\title{Calculate Q matrices from STRUCTURE v2.3.4 results}
\usage{
q_matrices(dir)
}
\arguments{
\item{dir}{The directory containing the STRUCTURE results.}
}
\value{
directory containing the q matrices files.
}
\description{
Calculate Q matrices from STRUCTURE v2.3.4 results
}
\examples{
q_matrices("./structure_res")
}
