% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{read_fasta}
\alias{read_fasta}
\title{Read FASTA files}
\usage{
read_fasta(zipped, directory)
}
\arguments{
\item{zipped}{The file path of the zipped file.}

\item{directory}{The directory to save the unpacked files. Default is the working directory.}
}
\value{
A DNA bin.
}
\description{
Read FASTA files
}
\keyword{internal}
