% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{read_msa_file}
\alias{read_msa_file}
\title{Read MSA files}
\usage{
read_msa_file(path, filename)
}
\arguments{
\item{path}{The file path of the alignment file.}

\item{filename}{The file name of input.}
}
\value{
The alignment.
}
\description{
Read MSA files
}
\keyword{internal}
