% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{revise_structure_file}
\alias{revise_structure_file}
\title{Convert dataframe to structure file}
\usage{
revise_structure_file(file, output.dir = ".", system = "Windows")
}
\arguments{
\item{file}{The dataframe containing sample, population, and genotype information.}

\item{to_str}{whether input file will be used to generate .str files.}
}
\value{
The file path to the .str file
}
\description{
Convert dataframe to structure file
}
\examples{
revise_structure_file(file = my_table, output.dir = ".", system = "Linux")
}
\seealso{
\code{\link[=convert_to_genind]{convert_to_genind()}} to convert sample, population, and genotype file to STRUCTURE v2.3.4-compatible format. Set 'to_str' to TRUE.
}
