% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_structure_analysis.R
\name{running_structure}
\alias{running_structure}
\title{Run STRUCTURE analysis}
\usage{
running_structure(
  input_file,
  k.range,
  num.k.rep,
  burnin,
  numreps,
  noadmix = FALSE,
  phased = FALSE,
  ploidy = NULL,
  linkage = FALSE,
  alpha_value = NULL,
  structure_path = structure_path,
  output_dir = dir
)
}
\description{
Run STRUCTURE analysis
}
\keyword{internal}
