% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_structure_analysis.R
\name{structureRead}
\alias{structureRead}
\title{Read STRUCTURE file
Sections commented out were revised.}
\usage{
structureRead(file, pops = NULL)
}
\description{
Read STRUCTURE file
Sections commented out were revised.
}
\keyword{internal}
