% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{subset_genind_pop}
\alias{subset_genind_pop}
\title{Subset populations for PCA recalculation}
\usage{
subset_genind_pop(fsnps_gen, pops)
}
\arguments{
\item{fsnps_gen}{The genind object containing sample, population, and genotype data.}

\item{pops}{The populations to be used for PCA recalculation.}
}
\value{
Genind object
}
\description{
Subset populations for PCA recalculation
}
\keyword{internal}
