% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{to_binary}
\alias{to_binary}
\title{Convert SNP genotypes to dosages}
\usage{
to_binary(df, markers = marker.file)
}
\arguments{
\item{df}{The dataframe of SNP calls.}

\item{markers}{The reference file containing metadata of samples with information on 1. rsID 2. REF 3. ALT.}
}
\value{
The dataframe of dosages per marker and samples.
}
\description{
Convert SNP genotypes to dosages
}
\examples{
to_binary(snp_calls, markers = marker_list)
}
