% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{to_snipper}
\alias{to_snipper}
\title{Convert dataframe to SNIPPER-analysis-compatible file}
\usage{
to_snipper(
  input,
  references,
  target.pop = TRUE,
  population.name = NULL,
  markers = snps
)
}
\arguments{
\item{input}{The dataframe with genotype and sample information.}

\item{references}{The dataframe containing population metadata of input. Sample names should match.}

\item{target.pop}{This indicates whether test data will be used. Otherwise, all data will be used included as a training set. Default is TRUE.}

\item{population.name}{The population to be used if test data will be used.}

\item{markers}{The number of markers in the dataset.}
}
\value{
A dataframe formatted for SNIPPER compatibility.
}
\description{
Convert dataframe to SNIPPER-analysis-compatible file
}
\examples{
to_snipper(input = df, references = pop_info, target.pop = FALSE, markers = 56)
}
