% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_structure_analysis.R
\name{to_structure}
\alias{to_structure}
\title{Convert genind object to STRUCTURE-compatible format}
\usage{
to_structure(genind_obj, include_pop = TRUE)
}
\description{
Convert genind object to STRUCTURE-compatible format
}
\keyword{internal}
