% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{unpack_input_file}
\alias{unpack_input_file}
\title{Unpack compressed files}
\usage{
unpack_input_file(files, output.dir = ".")
}
\arguments{
\item{files}{The file path of the zipped file.}

\item{output.dir}{The directory to save the unpacked files. Default is the working directory.}
}
\value{
File path of the directory where files are unpacked and the list of actual files.
}
\description{
Unpack compressed files
}
\keyword{internal}
