% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_structure_analysis.R
\name{utils.structure.evanno}
\alias{utils.structure.evanno}
\title{Parse matrix files
Sections commented out were revised.}
\usage{
utils.structure.evanno(sr, plot = TRUE)
}
\description{
Parse matrix files
Sections commented out were revised.
}
\keyword{internal}
