% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{vcf_to_csv}
\alias{vcf_to_csv}
\title{Convert VCF to CSV}
\usage{
vcf_to_csv(files, ref = NULL, output.dir = ".")
}
\arguments{
\item{files}{The file path of input file.}

\item{ref}{(optional) The dataframe containing metadata of samples.}

\item{output.dir}{(optional) The output directory to save input files if zipped. Default is working directory.}
}
\value{
The dataframe of merged metadata with genotype information.
}
\description{
Convert VCF to CSV
}
\examples{
vcf_to_csv('extracted_markers.vcf', ref = reference_file)
}
