% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{widen_genotype_file}
\alias{widen_genotype_file}
\title{Widen long genotype file}
\usage{
widen_genotype_file(files = files, population = NULL, output.dir = ".")
}
\arguments{
\item{files}{A zipped file containing long genotype files.}

\item{population}{(optional) The metadata of samples.}

\item{output.dir}{The directory to save unpacked files. Default is working directory.}
}
\value{
The dataframe of widened and merged genotype files.
}
\description{
Widen long genotype file
}
\examples{
widen_genotype_file(files = "zipped_files.zip", output.dir = "./new_dir")
}
