% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_statistics.R
\name{compute_hwe}
\alias{compute_hwe}
\title{Perform HWE test}
\usage{
compute_hwe(fsnps_gen)
}
\arguments{
\item{fsnps_gen}{The genind data containing genotype and population information.}
}
\value{
A list of HWE stats results.
}
\description{
Perform HWE test
}
\examples{
compute_hwe(fsnps_gen = my_genind)
}
