% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_pca.R
\name{compute_pca}
\alias{compute_pca}
\title{Find principal components}
\usage{
compute_pca(fsnps_gen, popinfo = TRUE)
}
\arguments{
\item{fsnps_gen}{The genind data containing genotype and population information.}

\item{popinfo}{To indicate if population metadata is present. Default is TRUE.}
}
\value{
A list of dataframes containing eigenvalues.
}
\description{
Find principal components
}
\examples{
compute_pca(genind_obj)
}
