% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{convert_to_genind}
\alias{convert_to_genind}
\title{Convert dataframes to genind object}
\usage{
convert_to_genind(file, to_str = FALSE, popinfo = TRUE)
}
\arguments{
\item{file}{A dataframe containing sample, population, and genotype information.}

\item{to_str}{To indicate whether input file will be used to generate .str files. Default is FALSE.}

\item{popinfo}{To indicate if population metadata is present. Default is TRUE.}
}
\value{
A genind object
}
\description{
Convert dataframes to genind object
}
\examples{
convert_to_genind(file = my_table, to_str = FALSE)
}
