% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_pca.R
\name{get_labels}
\alias{get_labels}
\title{Set up labels for PCA plotting}
\usage{
get_labels(fsnps_gen, use_default = TRUE, label_file = NULL, popinfo = TRUE)
}
\arguments{
\item{fsnps_gen}{The genind data containing genotype and population information.}

\item{use_default}{Indicates if default shapes and colors will be used. Default is TRUE.}

\item{label_file}{The file path to custom shapes and colors.}
}
\value{
A list of custom labels, shapes, and colors.
}
\description{
Set up labels for PCA plotting
}
\examples{
get_labels(genind_obj, use_default = FALSE, label_file = "custom.xlsx")
}
