% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{merge_plink2_files}
\alias{merge_plink2_files}
\title{Merge PLINK files}
\usage{
merge_plink2_files(base_prefix, merge_list, output_prefix)
}
\arguments{
\item{merge_list}{The list of PLINK 2.0 files.}

\item{output_prefix}{The prefix of the output (expected to be PLINK2.0).}
}
\value{
The file path of the output.
}
\description{
Merge PLINK files
}
\examples{
merge_plink2_files(merge_list = "plink_files.txt", output_prefix = "mergedFile")
}
