% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plots PCA}
\usage{
plot_pca(
  ind_coords,
  centroid,
  percent,
  labels_colors,
  width = 8,
  height = 8,
  pc_x = 1,
  pc_y = 2
)
}
\arguments{
\item{ind_coords}{The eigenvalues of the principal component.}

\item{centroid}{The the average eigenvalue per site/population.}

\item{percent}{The percentage of variance explained by the set x and y values.}

\item{labels_colors}{A list of custom shapes and colors.}

\item{width}{Output png's width.}

\item{height}{Output png's height.}

\item{pc_x}{Eigenvector to be plotted on the x-axis.}

\item{pc_y}{Eigenvector to be plotted on the y-axis.}
}
\value{
A PCA plot.
}
\description{
Plots PCA
}
\examples{
plot_pca(ind_coords, centroid, percent, labels_colors, width = 8, height = 8, pc_x = 1, pc_y = 2)
}
